FIX FOLDER RESTRUCTURATION (front):
Description Create a basic folder structure for the project start:
Notice: structure may vary in time, depends on features and reusability of somes code
Tree and (description)
.
├── doc_git.md
└── front (the folder contain source of the front)
    ├── app (the app that'll be binded to docker)
    │   ├── index.html (html source file)
    │   ├── package.json (app package)
    │   ├── public (public ressource folder)
    │   │   └── vite.svg
    │   ├── src	(Our workdir, we mainly modify code here)
    │   │   ├── App.tsx	 (First Compoment of our app)
    │   │   ├── custom.d.ts (Custom type declaration)
    │   │   ├── index.css (style called by main.tsx)
    │   │   ├── main.tsx (Entrypoint of our app, create the root element from index.html)
    │   │   ├── Views (Folder that contains views of app)
    │   │   │   └── BaseViewFromVitejs (A default View from vitejs installer)
    │   │   │       ├── assets	(asset of this view)
    │   │   │       │   └── react.svg
    │   │   │       ├── BaseViewFromViteJs.tsx (imported into App.tsx, has the same name as the current folder)
    │   │   │       ├── components (local components, when component is reused across view, will be in another folder)
    │   │   │       │   ├── CardBase.tsx
    │   │   │       │   ├── DocInfoBase.tsx
    │   │   │       │   ├── LogoBase.tsx
    │   │   │       │   └── TitleBase.tsx
    │   │   │       └── styles (style iof this view)
    │   │   │           └── Base.css
    │   │   └── vite-env.d.ts (type declaration)
    │   ├── tsconfig.json (typesctript config)
    │   ├── tsconfig.node.json (typescript config)
    │   └── vite.config.ts (vite config)
    ├── Dockerfile (the dockerfile for dev)
    └── Makefile (the Makefile for dev)
