+++ front/app/src/App.tsx
@@ -1,69 +1,10 @@
-import { useState } from "react";
-import reactLogo from "./assets/react.svg";
-import viteLogo from "/vite.svg";
-import "./App.css";
-
-const	LogoBase = () =>
-{
-	return (
-		<div>
-			<a href="https://vitejs.dev" target="_blank">
-				<img src={viteLogo} className="logo" alt="Vite logo" />
-			</a>
-			<a href="https://react.dev" target="_blank">
-				<img src={reactLogo} className="logo react" alt="React logo" />
-			</a>
-		</div>
-	);
-};
-
-const	TitleBase = () =>
-{
-	return (<h1>Vite + React</h1>);
-};
-
-const	CardBase = () =>
-{
-	const	[
-		count,
-		setCount
-	] = useState(0);
-
-	return (
-		<div className="card">
-			<button onClick={() =>
-			{
-				setCount((count) =>
-				{
-					return (count + 1);
-				});
-			}}>
-				count is {count}
-			</button>
-			<p>
-				Edit <code>src/App.tsx</code> and save to test HMR
-			</p>
-		</div>
-	);
-};
-
-const	DocInfoBase = () =>
-{
-	return (
-		<p className="read-the-docs">
-			Click on the Vite and React logos to learn more
-		</p>
-	);
-};
+import BaseViewFromViteJs from "./Views/BaseViewFromVitejs/BaseViewFromViteJs";
 
 const	App = () =>
 {
 	return (
 		<>
-			<LogoBase />
-			<TitleBase />
-			<CardBase />
-			<DocInfoBase />
+			<BaseViewFromViteJs />
 		</>
 	);
 };
