+++ front/app/src/Views/BaseViewFromVitejs/BaseViewFromViteJs.tsx
@@ -1,0 +1,21 @@
+
+import LogoBase from "./components/LogoBase";
+import TitleBase from "./components/TitleBase";
+import CardBase from "./components/CardBase";
+import DocInfoBase from "./components/DocInfoBase";
+
+import "./styles/Base.css";
+
+const	BaseViewFromViteJs = () =>
+{
+	return (
+		<>
+			<LogoBase />
+			<TitleBase />
+			<CardBase />
+			<DocInfoBase />
+		</>
+	);
+};
+
+export default BaseViewFromViteJs;
