+++ front/app/src/Views/BaseViewFromVitejs/components/DocInfoBase.tsx
@@ -1,0 +1,11 @@
+
+const	DocInfoBase = () =>
+{
+	return (
+		<p className="read-the-docs">
+			Click on the Vite and React logos to learn more
+		</p>
+	);
+};
+
+export default DocInfoBase;
