+++ front/app/src/Views/BaseViewFromVitejs/components/TitleBase.tsx
@@ -1,0 +1,7 @@
+
+const	TitleBase = () =>
+{
+	return (<h1>Vite + React</h1>);
+};
+
+export default TitleBase;
