+++ front/app/src/Views/BaseViewFromVitejs/styles/Base.css
